fix(amazonq): single click or select through arrow keys on review code issues does not focus code editor to line nor show issue details
